feat(79) : Taux d'affinité en Int et cas de formation sans taux

+++ app/back/src/main/fr/gouv/monprojetsup/recherche/application/dto/RecupererFormationReponseDTO.kt
@@ -27,7 +27,7 @@ data class RecupererFormationReponseDTO(
         val liens: List<LiensDTO>,
         val villes: List<String>,
         val metiers: List<MetierDetailleDTO>,
-        val tauxAffinite: Float?,
+        val tauxAffinite: Int?,
     ) {
         companion object {
             fun fromFicheFormation(ficheFormation: FicheFormation): FormationDetailleDTO {
